refactor most functions to view
