+++ lib/app.js
@@ -67,6 +67,8 @@ import ParticleSystemManager from './framework/particle/particle-system-manager'
 // other classes
 import { Keyframe, AnimationCurve } from './geom-utils/curve';
 
+import View from './platform/view';
+
 // register builtin asset loader
 registry.registerLoader('mesh', meshLoader);
 registry.registerLoader('material', materialLoader);
@@ -167,17 +169,10 @@ export default class App extends AppBase {
       poolSize: opts.poolSize || 100,
     });
     this.__initEventEmitter();
-
+    this._sys = new Sys();
     // sub-modules (canvas)
+    this._view = new View(this, canvas);
     this._canvas = canvas;
-    this._container = document.createElement('div');
-    if (this._canvas.parentNode) {
-      this._canvas.parentNode.insertBefore(this._container, this._canvas);
-    }
-    this._container.setAttribute('id', 'cocos3dContainer');
-    this._container.appendChild(this._canvas);
-    this._container.style.height = '100%';
-    this._availSize = vec2.new(300, 150);
 
     this._input = new Input(canvas, {
       lock: false,
@@ -201,7 +196,6 @@ export default class App extends AppBase {
     // sub-modules (engine)
     this._assetMng = new AssetMng(this);
     this._debugger = new Debugger(this);
-    this._sys = new Sys();
     this._configuration = new Configuration(this);
 
     // register builtin assets
@@ -227,14 +221,6 @@ export default class App extends AppBase {
     this._tickID = null;
     this._lasttime = 0;
     this._activeCamera = null;
-
-    window.addEventListener('resize', () => {
-      this.resize();
-    });
-
-    window.addEventListener('orientationchange', () => {
-      this._orientationChange();
-    });
   }
 
   get device() {
@@ -265,10 +251,6 @@ export default class App extends AppBase {
     return this._configuration;
   }
 
-  get availSize() {
-    return this._availSize;
-  }
-
   run() {
     if (!this._activeLevel) {
       console.warn('There is no level to run, please load it first');
@@ -283,15 +265,10 @@ export default class App extends AppBase {
       return;
     }
 
-    let bcr = this._container.parentElement.getBoundingClientRect();
-    vec2.set(this._availSize, bcr.width, bcr.height);
-    this._configuration.windowResize();
+    this._view.resize();
+    this._view.windowResize();
 
     this._input.resize();
-  }
-
-  _orientationChange() {
-    this._configuration.windowResize();
   }
 
   destroy() {
