+++ lib/platform/enums.js
@@ -1,0 +1,9 @@
+export default {
+  WEB_ORIENTATION_PORTRAIT: 0,
+  WEB_ORIENTATION_LANDSCAPE_LEFT: -90,
+  WEB_ORIENTATION_PORTRAIT_UPSIDE_DOWN: 180,
+  WEB_ORIENTATION_LANDSCAPE_RIGHT: 90,
+  ORIENTATION_PORTRAIT: 1,
+  ORIENTATION_LANDSCAPE: 2,
+  ORIENTATION_AUTO: 3,
+};
